Update only if needed

+++ test/update_catch.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
 
 MYDIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
-curl -s "https://raw.githubusercontent.com/jroivas/catch-me2-if-you-can/master/include/catch-me2.hpp" -o "${MYDIR}/catch.hpp"
+if [ ! -e "${MYDIR}/catch.hpp" ] || [ "${1:-}" == "--force" ]; then
+    curl -s "https://raw.githubusercontent.com/jroivas/catch-me2-if-you-can/master/include/catch-me2.hpp" -o "${MYDIR}/catch.hpp"
+fi
